move everything to the established decorator pattern
