+++ app/actors/iiif_print/actors/cleanup_file_sets_actor_decorator.rb
@@ -22,3 +22,4 @@ module IiifPrint
     end
   end
 end
+Hyrax::Actors::CleanupFileSetsActor.prepend(IiifPrint::Actors::CleanupFileSetsActorDecorator)
