+++ app/helpers/iiif_print/iiif_helper_decorator.rb
@@ -30,3 +30,4 @@ module IiifPrint
     end
   end
 end
+Hyrax::IiifHelper.prepend(IiifPrint::IiifHelperDecorator)
