+++ app/models/iiif_print/iiif_search_response_decorator.rb
@@ -38,3 +38,4 @@ module IiifPrint
     end
   end
 end
+::BlacklightIiifSearch::IiifSearchResponse.prepend(IiifPrint::IiifSearchResponseDecorator)
