+++ app/presenters/iiif_print/file_set_presenter_decorator.rb
@@ -9,3 +9,4 @@ module IiifPrint
     end
   end
 end
+Hyrax::FileSetPresenter.prepend(IiifPrint::FileSetPresenterDecorator)
