+++ app/renderers/hyrax/renderers/faceted_attribute_renderer_decorator.rb
@@ -16,3 +16,4 @@ module Hyrax
     end
   end
 end
+Hyrax::Renderers::FacetedAttributeRenderer.prepend(Hyrax::Renderers::FacetedAttributeRendererDecorator)
