+++ app/services/iiif_print/simple_schema_loader_decorator.rb
@@ -9,3 +9,4 @@ module IiifPrint
     end
   end
 end
+Hyrax::SimpleSchemaLoader.prepend(IiifPrint::SimpleSchemaLoaderDecorator)
