--- config/initializers/simple_schema_loader.rb
@@ -1,1 +1,0 @@
-"Hyrax::SimpleSchemaLoader".safe_constantize&.prepend(IiifPrint::SimpleSchemaLoaderDecorator)
