+++ lib/iiif_print/blacklight_iiif_search/annotation_decorator.rb
@@ -134,3 +134,4 @@ module IiifPrint
     end
   end
 end
+::BlacklightIiifSearch::IiifSearchAnnotation.prepend(IiifPrint::BlacklightIiifSearch::AnnotationDecorator)
